+++ packages/vsce/src/trees/CICSTree.ts
@@ -9,6 +9,9 @@
  *
  */
 
+import * as vscode from "vscode";
+import { FilterDescriptor } from "../utils/filterUtils";
+import { findSelectedNodes } from "../utils/commandUtils";
 import { getResource } from "@zowe/cics-for-zowe-sdk";
 import {
   Event,
@@ -17,9 +20,9 @@ import {
   ProviderResult,
   TreeDataProvider,
   TreeItem,
-  WebviewPanel,
   window,
-  commands
+  commands,
+  TreeView
 } from "vscode";
 import { PersistentStorage } from "../utils/PersistentStorage";
 import { InfoLoaded, ProfileManagement } from "../utils/profileManagement";
@@ -29,7 +32,7 @@ import { CICSRegionTree } from "./CICSRegionTree";
 import { CICSSessionTree } from "./CICSSessionTree";
 import * as https from "https";
 import { getIconPathInResources, missingSessionParameters, promptCredentials } from "../utils/profileUtils";
-import { Gui, imperative } from "@zowe/zowe-explorer-api";
+import { Gui, imperative, FileManagement, ZoweVsCodeExtension } from "@zowe/zowe-explorer-api";
 
 export class CICSTree implements TreeDataProvider<CICSSessionTree> {
   loadedProfiles: CICSSessionTree[] = [];
@@ -51,10 +54,10 @@ export class CICSTree implements TreeDataProvider<CICSSessionTree> {
   }
 
   /**
-   * Searches profiles stored in persistent storage, retrieves information for that profile from
-   * ZE's PorfilesCache API and then creates CICSSessionTrees with this information and adds
-   * these as children to the CICSTree (TreeDataProvider)
-   */
+     * Searches profiles stored in persistent storage, retrieves information for that profile from
+     * ZE's PorfilesCache API and then creates CICSSessionTrees with this information and adds
+     * these as children to the CICSTree (TreeDataProvider)
+     */
   public async loadStoredProfileNames() {
     const persistentStorage = new PersistentStorage("zowe.cics.persistent");
     await ProfileManagement.profilesCacheRefresh();
@@ -75,63 +78,118 @@ export class CICSTree implements TreeDataProvider<CICSSessionTree> {
   }
 
   /**
-   *
-   * Provides user with prompts and allows them to add a profile after clicking the '+' button
-   */
+     * Provides user with user actions and allows them to manage the selected profile
+     * @param treeview CICSTree View
+     * *@param node current selected node
+     */
+  async manageProfile(treeview: TreeView<any>, node: any) {
+    const allSelectedNodes = findSelectedNodes(treeview, CICSSessionTree, node);
+    if (!allSelectedNodes || !allSelectedNodes.length) {
+      window.showErrorMessage("No profile selected to delete");
+      return;
+    }
+    try {
+      const configInstance = await ProfileManagement.getConfigInstance();
+      if (configInstance.getTeamConfig().exists) {
+        const currentProfile = await ProfileManagement.getProfilesCache().getProfileFromConfig(
+          allSelectedNodes[allSelectedNodes.length - 1].label,
+        );
+
+        const deleteProfile: vscode.QuickPickItem = {
+          label: `$(trash) ${vscode.l10n.t("Delete Profile")}`,
+          description: vscode.l10n.t("Delete the selected Profile"),
+        };
+        const hideProfile: vscode.QuickPickItem = {
+          label: `$(eye-closed) ${vscode.l10n.t("Hide Profile")}`,
+          description: vscode.l10n.t("Hide profile name from tree view"),
+        };
+        const editProfile: vscode.QuickPickItem = {
+          label: `$(pencil) ${vscode.l10n.t("Edit Profile")}`,
+          description: vscode.l10n.t("Update profile connection information"),
+        };
+
+        const quickpick = Gui.createQuickPick();
+        const addProfilePlaceholder = "Choose user action for selected profile";
+        quickpick.items = [editProfile, hideProfile, deleteProfile];
+        quickpick.placeholder = addProfilePlaceholder;
+        quickpick.ignoreFocusOut = true;
+        quickpick.show();
+        const choice = await Gui.resolveQuickPick(quickpick);
+        quickpick.hide();
+        const debugMsg = vscode.l10n.t(`Profile selection has been cancelled.`);
+        if (!choice) {
+          Gui.showMessage(debugMsg);
+          return;
+        } else if (choice === hideProfile) {
+          this.hideZoweConfigFile(allSelectedNodes);
+          return;
+        } else if (choice === editProfile) {
+          for (const sessionTree of allSelectedNodes) {
+            await this.updateSession(sessionTree, configInstance);
+          }
+        } else {
+          const filePath = currentProfile.profLoc.osLoc[0];
+          await openConfigFile(filePath);
+          return;
+        }
+      }
+    } catch (error) {
+      window.showErrorMessage(
+        `Something went wrong while managing the profile - ${JSON.stringify(error, Object.getOwnPropertyNames(error)).replace(
+          /(\\n\t|\\n|\\t)/gm,
+          " ",
+        )}`,
+      );
+    }
+  }
+
+  /**
+     *
+     * Provides user with prompts and allows them to add a profile after clicking the '+' button
+     */
   async addProfile() {
     try {
       //const allCICSProfileNames = await ProfileManagement.getProfilesCache().getNamesForType('cics');
       const configInstance = await ProfileManagement.getConfigInstance();
-      const allCICSProfiles = (await ProfileManagement.getProfilesCache().getProfileInfo()).getAllProfiles("cics");
+      const profileInfo = await ProfileManagement.getProfilesCache().getProfileInfo();
+      const allCICSProfiles = profileInfo.getAllProfiles("cics");
       // const allCICSProfiles = await ProfileManagement.getProfilesCache().getProfiles('cics');
       const allCICSProfileNames: string[] = allCICSProfiles ? (allCICSProfiles.map((profile) => profile.profName) as unknown as [string]) : [];
       // No cics profiles needed beforhand for team config method
       if (configInstance.getTeamConfig().exists || allCICSProfileNames.length > 0) {
-        const profileNameToLoad = await window.showQuickPick(
-          [{ label: "\u270F Edit Team Configuration File" }].concat(
-            allCICSProfileNames
-              .filter((name) => {
-                for (const loadedProfile of this.loadedProfiles) {
-                  if (loadedProfile.label === name) {
-                    return false;
-                  }
-                }
-                return true;
-              })
-              .map((profileName) => {
-                return { label: profileName };
-              })
-          ),
-          {
-            ignoreFocusOut: true,
-            placeHolder: "Edit Team Configuration File",
-          }
-        );
-        if (profileNameToLoad) {
-          // If Create New CICS Profile option chosen
-          if (profileNameToLoad.label.includes("\u270F")) {
-            // get all profiles of all types including zosmf
-            const profiles = configInstance.getAllProfiles();
-            const currentProfile =
-                profiles.length > 0 ? await ProfileManagement.getProfilesCache().getProfileFromConfig(profiles[0].profName) : null;
-            const teamConfigFilePath = configInstance.getTeamConfig().opts.homeDir + "/zowe.config.json";
-            const filePath = currentProfile === null ? teamConfigFilePath : (currentProfile?.profLoc.osLoc?.[0] ?? teamConfigFilePath);
-            await openConfigFile(filePath);
-          } else {
-            let profileToLoad;
-            // TODO: Just use loadNamedProfile once the method is configured to v2 profiles
-            if (configInstance.getTeamConfig().exists) {
-              profileToLoad = await ProfileManagement.getProfilesCache().getLoadedProfConfig(profileNameToLoad.label); //ProfileManagement.getProfilesCache().loadNamedProfile(profileNameToLoad.label, 'cics');
-            } else {
-              await ProfileManagement.profilesCacheRefresh();
-              profileToLoad = ProfileManagement.getProfilesCache().loadNamedProfile(profileNameToLoad.label, "cics");
-            }
-            const newSessionTree = new CICSSessionTree(profileToLoad);
-            this.loadedProfiles.push(newSessionTree);
-            const persistentStorage = new PersistentStorage("zowe.cics.persistent");
-            await persistentStorage.addLoadedCICSProfile(profileNameToLoad.label);
-            this._onDidChangeTreeData.fire(undefined);
-          }
+        const createNewConfig = "Create a New Team Configuration File";
+        const editConfig = "Edit Team Configuration File";
+
+        const configPick = new FilterDescriptor("\uFF0B " + createNewConfig);
+        const configEdit = new FilterDescriptor("\u270F " + editConfig);
+        const items: vscode.QuickPickItem[] = [];
+
+        const profAllAttrs = profileInfo.getAllProfiles();
+        for (const pName of allCICSProfileNames) {
+          const osLocInfo = profileInfo.getOsLocInfo(profAllAttrs.find((p) => p.profName === pName));
+          items.push(new FilterDescriptor(this.getProfileIcon(osLocInfo)[0] + " " + pName));
+        }
+        const quickpick = Gui.createQuickPick();
+        const addProfilePlaceholder = vscode.l10n.t(`Choose "Create new..." to define or select a profile to add to the CICS tree`);
+        quickpick.items = [configPick, configEdit, ...items];
+        quickpick.placeholder = addProfilePlaceholder;
+        quickpick.ignoreFocusOut = true;
+        quickpick.show();
+        const choice = await Gui.resolveQuickPick(quickpick);
+        quickpick.hide();
+        const debugMsg = vscode.l10n.t(`Profile selection has been cancelled.`);
+        if (!choice) {
+          Gui.showMessage(debugMsg);
+          return;
+        } else if (choice === configPick) {
+          await this.createNewProfile();
+          return;
+        } else if (choice === configEdit) {
+          await this.editZoweConfigFile();
+          return;
+        } else {
+          await this.loadExistingProfile(choice.label);
+          return;
         }
       } else {
         //  Create New Profile Form should appear
@@ -144,12 +202,12 @@ export class CICSTree implements TreeDataProvider<CICSSessionTree> {
   }
 
   /**
-   *
-   * @param profile
-   * @param position number that's passed in when updating or expanding profile - needed
-   * to replace position of current CICSSessionTree.
-   * @param sessionTree current CICSSessionTree only passed in if expanding a profile
-   */
+     *
+     * @param profile
+     * @param position number that's passed in when updating or expanding profile - needed
+     * to replace position of current CICSSessionTree.
+     * @param sessionTree current CICSSessionTree only passed in if expanding a profile
+     */
   async loadProfile(profile?: imperative.IProfileLoaded, position?: number | undefined, sessionTree?: CICSSessionTree) {
     const persistentStorage = new PersistentStorage("zowe.cics.persistent");
     await persistentStorage.addLoadedCICSProfile(profile.name);
@@ -372,9 +430,9 @@ export class CICSTree implements TreeDataProvider<CICSSessionTree> {
   }
 
   /**
-   * Method for V1 profile configuration that provides UI for user to enter profile details
-   * and creates a profile.
-   */
+     * Method for V1 profile configuration that provides UI for user to enter profile details
+     * and creates a profile.
+     */
   async createNewProfile() {
     if (isTheia()) {
       const connnectionName = await Gui.showInputBox({
@@ -477,13 +535,57 @@ export class CICSTree implements TreeDataProvider<CICSSessionTree> {
         window.showErrorMessage(error);
       }
     } else {
-      //  Initialize new team configuration file
-      const response = await window.showQuickPick([{ label: "\uFF0B Create a New Team Configuration File" }], {
-        ignoreFocusOut: true,
-        placeHolder: "Create a New Team Configuration File",
-      });
-      if (response) {
-        commands.executeCommand("zowe.all.config.init");
+      commands.executeCommand("zowe.all.config.init");
+    }
+  }
+
+  /**
+     * Allows user to edit a configuration file based on global or project level context
+     */
+  async editZoweConfigFile() {
+    let rootPath = FileManagement.getZoweDir();
+    const workspaceDir = ZoweVsCodeExtension.workspaceRoot;
+    const choice = await this.getConfigLocationPrompt("edit");
+    if (choice === "global") {
+      await openConfigFile(rootPath + "/zowe.config.json");
+    } else if (choice === "project") {
+      rootPath = workspaceDir.uri.fsPath;
+      await openConfigFile(rootPath + "/zowe.config.json");
+    } else {
+      return;
+    }
+  }
+
+  /**
+     * Allows user to load an existing file from persistance area instead of creating a new profile.
+     * @param label name of the selected profile
+     */
+  async loadExistingProfile(label: string) {
+    label = label.split(/ (.*)/)[1];
+    const profileToLoad = await ProfileManagement.getProfilesCache().getLoadedProfConfig(label);
+    const newSessionTree = new CICSSessionTree(profileToLoad);
+    this.loadedProfiles.push(newSessionTree);
+    const persistentStorage = new PersistentStorage("zowe.cics.persistent");
+    await persistentStorage.addLoadedCICSProfile(label);
+    this._onDidChangeTreeData.fire(undefined);
+  }
+
+  /**
+     * Method for V1 profile configuration that provides UI for user to hide a selected profile.
+     * @param allSelectedNodes array of selected nodes
+     */
+  async hideZoweConfigFile(allSelectedNodes: any[]) {
+    for (const index in allSelectedNodes) {
+      try {
+        const currentNode = allSelectedNodes[parseInt(index)];
+        await this.removeSession(currentNode);
+      } catch (error) {
+        window.showErrorMessage(
+          `Something went wrong when hiding the profile - ${JSON.stringify(error, Object.getOwnPropertyNames(error)).replace(
+            /(\\n\t|\\n|\\t)/gm,
+            " ",
+          )}`,
+        );
       }
     }
   }
@@ -499,9 +601,9 @@ export class CICSTree implements TreeDataProvider<CICSSessionTree> {
   }
 
   /**
-   * Delete profile functionality for V1 profile configuration
-   * @param sessions
-   */
+     * Delete profile functionality for V1 profile configuration
+     * @param sessions
+     */
   async deleteSession(sessions: CICSSessionTree[]) {
     let answer;
     if (sessions.length === 1) {
@@ -553,25 +655,62 @@ export class CICSTree implements TreeDataProvider<CICSSessionTree> {
     }
   }
 
-  async updateSession(session: CICSSessionTree) {
+  /**
+     * Update profile functionality for V1 profile configuration
+     * @param session CICSSessions Tree
+     */
+  async updateSession(session: CICSSessionTree, configInstance: imperative.ProfileInfo) {
     await ProfileManagement.profilesCacheRefresh();
     const profileCache = await ProfileManagement.getProfilesCache();
     const profileToUpdate = profileCache.loadNamedProfile(session.label?.toString()!, "cics");
     const currentProfile = await profileCache.getProfileFromConfig(profileToUpdate.name);
-    await this.updateSessionHelper(currentProfile);
+    const teamConfigFilePath = configInstance.getTeamConfig().opts.homeDir + "/zowe.config.json";
+    const filePath = currentProfile?.profLoc.osLoc?.[0] ?? teamConfigFilePath;
+    await openConfigFile(filePath);
   }
 
-  async updateSessionHelper(profile: imperative.IProfAttrs) {
-    const response = await window.showQuickPick([{ label: "\u270F Edit CICS Profile" }], {
-      ignoreFocusOut: true,
-      placeHolder: "Create a New Team Configuration File",
-    });
-    if (response) {
-      const configInstance = await ProfileManagement.getConfigInstance();
-      const teamConfigFilePath = configInstance.getTeamConfig().opts.homeDir + "/zowe.config.json";
-      const filePath = profile?.profLoc.osLoc?.[0] ?? teamConfigFilePath;
-      await openConfigFile(filePath);
+  /**
+     * Method for V1 profile configuration that returns the context of a configuration file.
+     * @param action string create or edit
+     */
+  private async getConfigLocationPrompt(action: string): Promise<string> {
+    let placeHolderText: string;
+    if (action === "create") {
+      placeHolderText = vscode.l10n.t("Select the location where the config file will be initialized");
+    } else {
+      placeHolderText = vscode.l10n.t("Select the location of the config file to edit");
     }
+    const quickPickOptions: vscode.QuickPickOptions = {
+      placeHolder: placeHolderText,
+      ignoreFocusOut: true,
+      canPickMany: false,
+    };
+    const globalText = vscode.l10n.t("Global: in the Zowe home directory");
+    const projectText = vscode.l10n.t("Project: in the current working directory");
+    const location = await Gui.showQuickPick([globalText, projectText], quickPickOptions);
+    // call check for existing and prompt here
+    switch (location) {
+      case globalText:
+        return "global";
+      case projectText:
+        return "project";
+    }
+  }
+
+  /**
+     * Method that returns the icon based on global or local context.
+     * @param osLocInfo physical location of of profile on OS
+     */
+  private getProfileIcon(osLocInfo: imperative.IProfLocOsLoc[]): string[] {
+    const ret: string[] = [];
+    for (const loc of osLocInfo ?? []) {
+      if (loc.global) {
+        ret.push("$(home)");
+      } else {
+        ret.push("$(folder)");
+      }
+    }
+    return ret;
   }
 
   getTreeItem(element: CICSSessionTree): TreeItem | Thenable<TreeItem> {