+++ packages/vsce/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to the "cics-extension-for-zowe" extension will be documented in this file.
 
+## Recent Changes
+- BugFix: Now the 'Show SIT Parameters' menu option is grayed out for inactive regions. [#167](https://github.com/zowe/cics-for-zowe-client/issues/167)
+
 ## `3.2.0`
 
 - BugFix: Initializing new Team Configuration File. [#33](https://github.com/zowe/cics-for-zowe-client/issues/33)